requirements.txt
apache-airflow==2.11.0  # UPGRADED
pandas==2.1.3
pymongo==4.6.0
sqlalchemy    # NO LONGER CONFLICTS!
psycopg2-binary==2.9.9
pydantic==2.5.0
datasets==2.14.7
requests==2.31.0
matplotlib==3.8.2
python-dotenv==1.0.0
pytest==7.4.3
apache-airflow-providers-openlineage==1.10.0